# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/bookings/": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Bookings",
        "operationId": "get_bookings_bookings__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "show_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Show Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingOut"
                  },
                  "title": "Response Get Bookings Bookings  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Create Booking",
        "operationId": "create_booking_bookings__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/{booking_id}/logs": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Booking Logs",
        "operationId": "get_booking_logs_bookings__booking_id__logs_get",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/cancel/{booking_id}": {
      "put": {
        "tags": [
          "Bookings"
        ],
        "summary": "Delete Booking",
        "operationId": "delete_booking_bookings_cancel__booking_id__put",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bookings/{booking_id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Booking",
        "operationId": "get_booking_bookings__booking_id__get",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookedFoodOut": {
        "properties": {
          "booking_id": {
            "type": "integer",
            "title": "Booking Id"
          },
          "food_id": {
            "type": "integer",
            "title": "Food Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          },
          "gst_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gst Id",
            "description": "FK to GST"
          },
          "booked_food_id": {
            "type": "integer",
            "title": "Booked Food Id"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "food_id",
          "quantity",
          "unit_price",
          "booked_food_id"
        ],
        "title": "BookedFoodOut"
      },
      "BookedSeatOut": {
        "properties": {
          "booking_id": {
            "type": "integer",
            "title": "Booking Id"
          },
          "seat_id": {
            "type": "integer",
            "title": "Seat Id"
          },
          "show_id": {
            "type": "integer",
            "title": "Show Id"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "gst_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gst Id",
            "description": "FK to GST"
          },
          "booked_seat_id": {
            "type": "integer",
            "title": "Booked Seat Id"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "seat_id",
          "show_id",
          "price",
          "booked_seat_id"
        ],
        "title": "BookedSeatOut"
      },
      "BookingCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "show_id": {
            "type": "integer",
            "title": "Show Id"
          },
          "booking_reference": {
            "type": "string",
            "maxLength": 20,
            "title": "Booking Reference"
          },
          "booking_status": {
            "$ref": "#/components/schemas/BookingStatus",
            "default": "PENDING"
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id",
            "description": "FK to PAYMENT"
          },
          "discount_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Id",
            "description": "FK to DISCOUNTS"
          },
          "booking_time": {
            "type": "string",
            "format": "date-time",
            "title": "Booking Time"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "seats": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seats",
            "description": "List of seat IDs booked"
          },
          "foods": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foods",
            "description": "List of food IDs ordered"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "show_id",
          "booking_reference",
          "booking_time",
          "amount"
        ],
        "title": "BookingCreate"
      },
      "BookingOut": {
        "properties": {
          "booking_id": {
            "type": "integer",
            "title": "Booking Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "show_id": {
            "type": "integer",
            "title": "Show Id"
          },
          "booking_reference": {
            "type": "string",
            "title": "Booking Reference"
          },
          "booking_status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          },
          "discount_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Id"
          },
          "booking_time": {
            "type": "string",
            "format": "date-time",
            "title": "Booking Time"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "seats": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BookedSeatOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seats"
          },
          "foods": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BookedFoodOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foods"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "user_id",
          "show_id",
          "booking_reference",
          "booking_status",
          "booking_time",
          "amount"
        ],
        "title": "BookingOut"
      },
      "BookingStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONFIRMED",
          "CANCELLED"
        ],
        "title": "BookingStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "JWTBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}